add fallback for net traffic sampler
Adds a fallback network traffic sampler that reads from sysfs when
BPF is disabled or not available.
