+++ src/samplers/network/linux/traffic/mod.rs
@@ -1,5 +1,4 @@
-use crate::common::Nop;
-use crate::samplers::network::*;
+use crate::samplers::network::linux::*;
 
 const NAME: &str = "network_traffic";
 
@@ -16,12 +15,34 @@ fn init(config: &Config) -> Box<dyn Sampler> {
     if let Ok(s) = NetworkTraffic::new(config) {
         Box::new(s)
     } else {
-        Box::new(Nop {})
+        let metrics = vec![
+            (&NETWORK_RX_BYTES, "rx_bytes"),
+            (&NETWORK_RX_PACKETS, "rx_packets"),
+            (&NETWORK_TX_BYTES, "tx_bytes"),
+            (&NETWORK_TX_PACKETS, "tx_packets"),
+        ];
+
+        if let Ok(s) = SysfsNetSampler::new(config, NAME, metrics) {
+            Box::new(s)
+        } else {
+            Box::new(Nop::new(config))
+        }
     }
 }
 
 #[cfg(not(feature = "bpf"))]
 #[distributed_slice(NETWORK_SAMPLERS)]
 fn init(config: &Config) -> Box<dyn Sampler> {
-    Box::new(Nop {})
+    let metrics = vec![
+        (&NETWORK_RX_BYTES, "rx_bytes"),
+        (&NETWORK_RX_PACKETS, "rx_packets"),
+        (&NETWORK_TX_BYTES, "tx_bytes"),
+        (&NETWORK_TX_PACKETS, "tx_packets"),
+    ];
+
+    if let Ok(s) = SysfsNetSampler::new(config, NAME, metrics) {
+        Box::new(s)
+    } else {
+        Box::new(Nop::new(config))
+    }
 }
